11Rotational Force 3Zeros Bug Fix
That's got to be the stupidest (but tutorial induced) bug I've had in 20 years! #TrustNoOne (Especially Tutorials) Apparently the tutorial missed saying the angular member variables should be set to zero ... and I have to say doing it in header rather than an explicit initialisation section is pretty bad practice (unless it is a universal constant like pi). Here is it with the 3 zeros added (sigh!): https://clipchamp.com/watch/GSZvtECfrD2